ObjHits_RecordPositionHit 98.75->100 (objlib 99.941->99.971): retail cross-TU param shape - the DEFINITION takes s8 priority (bare stb of the param, no cast-CSE'd extsb store) while retail callers compiled against the int view; (int) head-compare cast keeps the two extsb sites un-CSE'd, objhits.c passes the s8 field uncast (retail's bare lbz r5), player.c's two hv sites keep the int view via (int)hv (retail's li+extsb; fn_802AD2F4 stays 0-region). objhits.o back to baseline byte-exact, player unit unchanged 99.709

+++ include/main/objhits.h
@@ -313,7 +313,7 @@ int ObjHits_IsObjectEnabled(ObjAnimComponent* obj);
 void ObjHits_SyncObjectPosition(GameObject* obj);
 void ObjHits_RefreshObjectState(GameObject* obj);
 void ObjHits_AddContactObject(GameObject* obj, GameObject* contactObj);
-int ObjHits_RecordPositionHit(GameObject* obj, GameObject* hitObj, int priority, int hitVolume, s8 sphereIndex,
+int ObjHits_RecordPositionHit(GameObject* obj, GameObject* hitObj, s8 priority, int hitVolume, s8 sphereIndex,
                               f32 hitPosX, f32 hitPosY, f32 hitPosZ);
 int ObjHits_RecordObjectHit(GameObject* obj, GameObject* hitObj, s8 priority, int hitVolume, s8 sphereIndex);
 int ObjHits_GetPriorityHitWithPosition(GameObject* obj, int* outHitObject, int* outSphereIndex, u32* outHitVolume,
